requirements.txt
docling
python-telegram-bot==20.*
pytest
pytest-asyncio
pytest-mock
